updated routes and index.js files
