+++ package.json
@@ -1,39 +1,34 @@
 {
-  "name": "Project2",
+  "name": "dinnerAndAMovie",
   "version": "1.0.0",
   "description": "A fullstack project brought to you by my teammates and I",
-  "main": "server.js",
+  "main": "index.js",
   "scripts": {
-    "build-css": "tailwindcss build Develop/src/style.css -o public/style.css", 
-    "start": "node server.js"
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/DaronSchmit/Project2.git"
+    "url": "git+https://github.com/DaronSchmit/DinnerAndAMovie.git"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "bugs": {
-    "url": "https://github.com/DaronSchmit/Project2/issues"
+    "url": "https://github.com/DaronSchmit/DinnerAndAMovie/issues"
   },
-  "homepage": "https://github.com/DaronSchmit/Project2#readme",
+  "homepage": "https://github.com/DaronSchmit/DinnerAndAMovie#readme",
   "dependencies": {
-    "autoprefixer": "^10.2.1",
+    "axios": "^0.21.1",
     "bcryptjs": "^2.4.3",
+    "dotenv": "^8.2.0",
+    "eslinter": "^3.2.1",
     "express": "^4.17.1",
+    "express-handlebars": "^5.2.0",
     "express-session": "^1.17.1",
     "mysql2": "^2.2.5",
     "passport": "^0.4.1",
-    "postcss": "^8.2.4",
-    "tailwindcss": "^2.0.2"
     "passport-local": "^1.0.0",
-    "sequelize": "^6.3.5",
-    "dotenv": "^8.2.0"
-  },
-  "devDependencies": {
-    "eslint": "^7.18.0",
-    "eslint-config-airbnb-base": "^14.2.1",
-    "eslint-plugin-import": "^2.22.1"
+    "sequelize": "^6.4.0",
+    "travis": "^0.1.1"
   }
 }
