+++ public/assets/js/favorites.js
