+++ routes/htmlRoutes/index.js
@@ -1,0 +1,7 @@
+const router = express.Router();
+
+const htmlRoutes = require("./html-routes");
+
+router.use("/", htmlRoutes);
+
+module.router = router;
